Push update for voice controller
Push intial codebase for voice controller logic. I have not fully implemented this because I switch to a new voice AI system, but it is pretty much done absent some testing.

+ users can define voice actions with a list of pre-defined functions.
+ users can define key actions with any key that is deemed pressable by robot.js.
+ test code for a potential new feature allowing express to detect the host, which will remove the need to configure this.
+ add two new post methods to meet the server-side need for voice commands.
+ add utility files to keep the main controller clean and primarily routes.

updates to come will include full rollout of voice functionality, as well as a detailed explanation of configuring and setting up the voice controller.